requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pymongo==4.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0
